pyproject.toml
[project]
name = "agent-second-brain"
version = "1.0.0"
description = "Voice-first personal assistant for Telegram with Obsidian and Todoist integration"
readme = "README.md"
authors = [
    { name = "smixs", email = "mixshima@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0",
    "deepgram-sdk",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "httpx",
    "pydantic>=2.0",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "d_brain"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
